# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Loan CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # TEAMS
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients (filtered by role automatically)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ativo, em_cobranca, quitado]
        - name: vendedor_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client and auto-generate installments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Client created with installments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client with full details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}/risk-level:
    patch:
      operationId: updateClientRiskLevel
      tags: [clients]
      summary: Manually set or clear client risk level
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskLevelInput"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  # INSTALLMENTS
...
  /clients/{id}/renovacao:
    post:
      operationId: renovacaoClient
      tags: [installments]
      summary: Generate new installments for renewal/agreement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenovacaoInput"
      responses:
        "200":
          description: Renewal registered with new installments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  # CLIENT EVENTS
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List client events (global timeline)
      parameters:
        - name: client_id
          in: query
          schema:
            type: integer
        - name: tipo
          in: query
          schema:
            type: string
            enum: [venda_fechada, parcela_paga, atraso, renovacao, follow_up, anotacao]
        - name: user_id
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientEvent"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a client event (follow_up, anotacao, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientEvent"

  # TASKS
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, senha]
      properties:
        email:
          type: string
        senha:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
        active_team_id:
          type: ["integer", "null"]
        needs_team:
          type: boolean

    # TEAMS
...
    User:
      type: object
      required: [id, nome, email]
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        papel:
          type: ["string", "null"]
          enum: [vendedor, cobrador, lider, null]
        active_team_id:
          type: ["integer", "null"]
        active_team:
          type: ["object", "null"]
        needs_team:
          type: boolean
        createdAt:
          type: string
...
    Client:
      type: object
      required: [id, nome, telefone, vendedor_id, valor_contrato, parcelas, dia_vencimento, status]
      properties:
        id:
          type: integer
        nome:
          type: string
        telefone:
          type: string
        vendedor_id:
          type: integer
        vendedor_nome:
          type: ["string", "null"]
        valor_contrato:
          type: number
        parcelas:
          type: integer
        dia_vencimento:
          type: integer
        status:
          type: string
          enum: [ativo, em_cobranca, quitado]
        risk_level:
          type: ["string", "null"]
          enum: [atencao, risco, critico, null]
        createdAt:
          type: string
    ClientDetail:
      type: object
      required: [id, nome, telefone, vendedor_id, valor_contrato, parcelas, dia_vencimento, status, installments, events]
      properties:
        id:
          type: integer
        nome:
          type: string
        telefone:
          type: string
        vendedor_id:
          type: integer
        vendedor_nome:
          type: ["string", "null"]
        valor_contrato:
          type: number
        parcelas:
          type: integer
        dia_vencimento:
          type: integer
        status:
          type: string
          enum: [ativo, em_cobranca, quitado]
        risk_level:
          type: ["string", "null"]
          enum: [atencao, risco, critico, null]
        createdAt:
          type: string
        installments:
          type: array
          items:
            $ref: "#/components/schemas/Installment"
        events:
          type: array
          items:
            $ref: "#/components/schemas/ClientEvent"
    RiskLevelInput:
      type: object
      required: [risk_level]
      properties:
        risk_level:
          type: ["string", "null"]
          enum: [atencao, risco, critico, null]
    ClientInput:
      type: object
      required: [nome, telefone, valor_contrato, parcelas, dia_vencimento]
      properties:
        nome:
          type: string
        telefone:
          type: string
        valor_contrato:
          type: number
        parcelas:
          type: integer
        dia_vencimento:
          type: integer
        vendedor_id:
          type: ["integer", "null"]
    ClientUpdate:
      type: object
      properties:
        nome:
          type: string
        telefone:
          type: string
        status:
          type: string
          enum: [ativo, em_cobranca, quitado]
        valor_contrato:
          type: number

    # INSTALLMENTS
    Installment:
      type: object
      required: [id, client_id, numero_parcela, valor, vencimento, status]
      properties:
        id:
          type: integer
        client_id:
          type: integer
        client_nome:
          type: ["string", "null"]
        vendedor_id:
          type: ["integer", "null"]
        numero_parcela:
          type: integer
        valor:
          type: number
        vencimento:
          type: string
        status:
          type: string
          enum: [pendente, pago, atrasado]
        pago_em:
          type: ["string", "null"]
...
    RenovacaoInput:
      type: object
      required: [novas_parcelas, novo_valor, novo_dia_vencimento]
      properties:
        novas_parcelas:
          type: integer
        novo_valor:
          type: number
        novo_dia_vencimento:
          type: integer
        observacao:
          type: string

    # EVENTS
    ClientEvent:
      type: object
      required: [id, client_id, tipo, user_id, data]
      properties:
        id:
          type: integer
        client_id:
          type: integer
        client_nome:
          type: ["string", "null"]
        tipo:
          type: string
          enum: [venda_fechada, parcela_paga, atraso, renovacao, follow_up, anotacao]
        user_id:
          type: integer
        user_nome:
          type: ["string", "null"]
        observacao:
          type: ["string", "null"]
        data:
          type: string
    EventInput:
      type: object
      required: [client_id, tipo]
      properties:
        client_id:
          type: integer
        tipo:
          type: string
          enum: [venda_fechada, parcela_paga, atraso, renovacao, follow_up, anotacao]
        observacao:
          type: string

    # TASKS
...